Fix SSR error with third-party-libs #1151
